requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0

# Deep Learning
tensorflow==2.13.0
keras==2.13.1

# Model Optimization and Deployment
onnx==1.14.0
onnxruntime==1.15.1
tf2onnx==1.15.1

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Dashboard and Monitoring
dash==2.11.1
dash-bootstrap-components==1.4.2

# Time Series Analysis
statsmodels==0.14.0

# Configuration and Utilities
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging and Monitoring
colorlog==6.7.0

# Data Generation and Testing
faker==19.2.0

# Progress Bars
tqdm==4.65.0

# Job Scheduling (for periodic tasks)
schedule==1.2.0

# API and Communication
requests==2.31.0
flask>=1.0.4,<2.3.0

# Model Serialization
joblib==1.3.1

# Date and Time
python-dateutil==2.8.2

# Memory Profiling (Optional)
memory-profiler==0.61.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# ========== NEW IMPROVEMENTS ==========

# REST API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database (TimescaleDB/PostgreSQL)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# MQTT for IoT Streaming
paho-mqtt==1.6.1

# MLflow for Experiment Tracking
mlflow==2.8.1

# Model Explainability
shap==0.43.0
lime==0.2.0.1

# Hyperparameter Optimization
optuna==3.4.0

# Prometheus Metrics
prometheus-client==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# WebSocket Support
websockets==12.0
flask-socketio==5.3.5
python-socketio==5.10.0

# Advanced Signal Processing
pywavelets==1.5.0

# Task Queue (for background jobs)
celery==5.3.4
redis==5.0.1

# Additional Utilities
httpx==0.25.1
aiofiles==23.2.1
